Checkpoint commit: mostly reliable?
iPod operation appears reliable on my workbench and in the BMW.  Try not to
mess with anything and see if it continues to be reliable.

Based on what I'm seeing on the oscilloscope, the serial timing isn't very
accurate.  Bit widths at 19200 should be 52.08µs but are 56µs instead, which
makes the actual baud rate 17857, or about 7% off.  The iPod seems to adapt to
this perfectly, but it still isn't correct.